Instance Manager

+++ php/Instance.php
@@ -7,21 +7,61 @@
 
 require_once 'DataBase.php';
 require_once 'XML.php';
+require_once 'Session.php';
 
+if (!isset($_SESSION))
+    session_start();
 
 class Instance {
     public function __construct() {
         $option = filter_input(INPUT_POST, "option");
         switch ($option)
         {
-            case 'GetInstances': $this->GetInstances(); break;
+            case 'getInstances': $this->getInstances(); break;
             case "DeleteInstance": $this->DeleteInstance(); break;
             case 'buildNewInstance': $this->buildNewInstance(); break;
         }
     }
     
     private function buildNewInstance(){
-        var_dump($_POST);
+        $DB = new DataBase();
+        $instanceName = filter_input(INPUT_POST, "instanceName");
+        $userName = filter_input(INPUT_POST, "userName");
+        
+        $RoutFile = dirname(getcwd());
+        
+        $createSchema = "CREATE DATABASE $instanceName 
+                        DEFAULT CHARACTER SET utf8
+                        DEFAULT COLLATE utf8_general_ci";
+        
+        if(file_exists("$RoutFile/Estructuras/$instanceName"))
+                return XML::XMLReponse("Error", 0, "La instancia ingresada ya existe.");
+        
+        if(($result = $DB->ConsultaQuery("", $createSchema))!=1)
+                return XML::XMLReponse ("Error", 0, "<b>Error</b> al intentar crear la instancia $instanceName.<br>Detalles:<br>$result" );
+        
+        if(($resultBuildInstanceControl = $DB->CreateCSDocsControl($instanceName))!=1){
+                $DB->ConsultaQuery("", "DROP DATABASE IF EXISTS $instanceName");
+                return XML::XMLReponse ("Error", 0, "<b>Error</b> al crear la estructura de control de la instancia <b>$instanceName</b>. $resultBuildInstanceControl");
+        }
+        if($resultBuildInstanceControl == 0){
+            $DB->ConsultaQuery("", "DROP DATABASE IF EXISTS $instanceName");
+        }
+        
+        $InsertInstance = "INSERT INTO instancias (NombreInstancia, fechaCreacion, usuarioCreador) VALUES ('$instanceName', 'now()', '$userName')";
+        
+        if(($resultInsert = $DB->ConsultaQuery("cs-docs", $InsertInstance))!=1){
+            $DB->ConsultaQuery("", "DROP DATABASE IF EXISTS $instanceName");
+            return XML::XMLReponse("Error", 0, "<b>Error</b> al intentar registrar la instancia. <br>Detalles:<br>$resultInsert");
+        }
+        
+        mkdir("$RoutFile/Estructuras/$instanceName", 0777);
+        
+        /* Archivo que almacena la configuración de la estrucutra de usuarios, empresas, repositorios, etc */
+        touch("$RoutFile/Configuracion/$instanceName.ini");
+       
+        return XML::XMLReponse("newInstanceBuilded", 1, "Instancia construida con éxito.");
+        
     }
     
     private function DeleteInstance()
@@ -31,7 +71,21 @@ class Instance {
         $UserName = filter_input(INPUT_POST, "UserName");
         $IdInstance = filter_input(INPUT_POST, "IdInstance");
         $InstanceName = filter_input(INPUT_POST, "InstanceName");
-        $RoutFile = filter_input(INPUT_SERVER, "DOCUMENT_ROOT"); /* /var/services/web */
+        
+        $idSession = Session::getIdSession();
+
+        if($idSession == null)
+                return XML::XMLReponse ("Error", 0, "Tree:No existe una sesión activa, por favor vuelva a iniciar sesión");
+            
+        $userData = Session::getSessionParameters();
+        
+        /* Sí el usuario inicio sesión en la misma instancia que va a eliminar, se destruye la sesión */
+        if(is_array($userData)){
+            if(isset($userData['dataBaseName']))
+                Session::destroySession ();
+        }
+        
+        $RoutFile = dirname(getcwd());
         
         $QueryDrop = "DROP DATABASE IF EXISTS $InstanceName";
         if(($Result = $DB->ConsultaQuery("cs-docs", $QueryDrop))!=1)
@@ -79,12 +133,12 @@ class Instance {
         
         if(file_exists("$RoutFile/Log/$InstanceName"))
             exec ("rm -R $RoutFile/Log/$InstanceName");
-        
+                
         XML::XMLReponse("DeleteInstance", 1, "Instancia $InstanceName eliminada");
         
     }
     
-    private function GetInstances()
+    private function getInstances()
     {        
         $instances = $this->getInstancesArray();
